# Pipeline chunk files benchmark