Upgrading to [moov-io/ach v1.35.0](https://github.com/moov-io/ach/releases/tag/v1.35.0) as well, which has some performance improvements. 

Without the moov-io/ach upgrade this benchmark reports 
```
Benchmark_chunkFilesTogether
    merging_test.go:172: merged 10000 files
Benchmark_chunkFilesTogether/run
    merging_test.go:191: 3 files after merge
Benchmark_chunkFilesTogether/run-16         	       1	7218584114 ns/op	230712568 B/op	  871031 allocs/op
PASS
ok  	github.com/moov-io/achgateway/internal/pipeline	14.087s
```

After upgrading moov-io/ach
```
Benchmark_chunkFilesTogether
    merging_test.go:172: merged 10000 files
Benchmark_chunkFilesTogether/run
    merging_test.go:191: 3 files after merge
Benchmark_chunkFilesTogether/run-16         	       1	7234198724 ns/op	229498080 B/op	  700987 allocs/op
PASS
ok  	github.com/moov-io/achgateway/internal/pipeline	14.208s
```

There's still a lot of room for improvement in the merge files routine, but adding Filename to `FileUploaded` shouldn't hinder uploads.